lib/requirements.txt
annotated-types==0.7.0
apturl==0.5.2
attrs==21.2.0
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
dell-recovery==0.0.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
Flask==3.1.0
Flask-Cors==5.0.0
future==0.18.2
google-ai-generativelanguage==0.6.13
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai @ git+https://github.com/google/generative-ai-python@b8772ed1424a080911151b354764d76a0e7af2af
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gTTS==2.5.4
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
playsound==1.3.0
proto-plus==1.25.0
protobuf==5.29.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.3
pydantic_core==2.27.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.32.3
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.3
systemd-python==234
texttable==1.6.4
tqdm==4.67.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
